requirements.txt
amqp==5.0.2
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
decorator==4.4.2
Django==3.1.3
django-ace==1.0.10
django-celery-beat==2.1.0
django-login-required-middleware==0.5.0
django-mathfilters==1.0.0
django-timezone-field==4.0
djangorestframework==3.12.2
djangorestframework-datatables==0.5.2
dnspython==2.0.0
exrex==0.10.5
filelock==3.0.12
fire==0.3.1
future==0.18.2
gunicorn==20.0.4
idna==2.10
kombu==5.0.2
loguru==0.5.3
prompt-toolkit==3.0.8
psycopg2==2.8.6
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-file==1.5.1
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
tenacity==6.2.0
termcolor==1.1.0
tldextract==3.1.0
tqdm==4.51.0
treelib==1.6.1
urllib3==1.25.11
validators==0.18.1
vine==5.0.0
wcwidth==0.2.5
APScheduler==3.6.3
cachetools==4.2.2
python-telegram-bot==13.7
tzlocal==2.1
tornado==6.1
